server/requirements.txt
# FastAPI server dependencies
#
# Install:
#   python -m pip install -r requirements.txt
#
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
python-jose[cryptography]>=3.3
python-multipart>=0.0.7
pydantic>=2.6
email-validator>=2.0
